project-templates/django-2-api/requirements/local.txt
-r base.txt

django-debug-toolbar==1.9.1

# required by django-extension
Werkzeug==0.12.2
django-extensions==1.9.9
